+++ test-tests/shared/src/test/scala/zio/test/ReportingTestUtils.scala
@@ -6,9 +6,9 @@ import zio.clock.Clock
 import zio.test.Assertion.{ equalTo, isGreaterThan, isLessThan, isRight, isSome, not }
 import zio.test.environment.{ testEnvironment, TestClock, TestConsole, TestEnvironment }
 import zio.test.mock.Expectation._
-import zio.test.mock.ModuleMock
 import zio.test.mock.internal.InvalidCall._
 import zio.test.mock.internal.MockException._
+import zio.test.mock.module.ModuleMock
 import zio.{ Cause, Layer, ZIO }
 
 object ReportingTestUtils {
@@ -181,10 +181,10 @@ object ReportingTestUtils {
   val mock1Expected = Vector(
     expectedFailure("Invalid call"),
     withOffset(2)(s"${red("- could not find a matching expectation")}\n"),
-    withOffset(4)(s"${red("- zio.test.mock.ModuleMock.Command called with invalid arguments")}\n"),
+    withOffset(4)(s"${red("- zio.test.mock.module.ModuleMock.Command called with invalid arguments")}\n"),
     withOffset(6)(s"${blue("2")} did not satisfy ${cyan("equalTo(1)")}\n"),
-    withOffset(4)(s"${red("- invalid call to zio.test.mock.ModuleMock.SingleParam")}\n"),
-    withOffset(6)(s"expected zio.test.mock.ModuleMock.Command with arguments ${cyan("equalTo(1)")}\n")
+    withOffset(4)(s"${red("- invalid call to zio.test.mock.module.ModuleMock.SingleParam")}\n"),
+    withOffset(6)(s"expected zio.test.mock.module.ModuleMock.Command with arguments ${cyan("equalTo(1)")}\n")
   )
 
   val mock2 = zio.test.test("Unsatisfied expectations") {
@@ -198,8 +198,8 @@ object ReportingTestUtils {
     expectedFailure("Unsatisfied expectations"),
     withOffset(2)(s"${red("- unsatisfied expectations")}\n"),
     withOffset(4)(s"in sequential order\n"),
-    withOffset(6)(s"""zio.test.mock.ModuleMock.SingleParam with arguments ${cyan("equalTo(2)")}\n"""),
-    withOffset(6)(s"""zio.test.mock.ModuleMock.SingleParam with arguments ${cyan("equalTo(3)")}\n""")
+    withOffset(6)(s"""zio.test.mock.module.ModuleMock.SingleParam with arguments ${cyan("equalTo(2)")}\n"""),
+    withOffset(6)(s"""zio.test.mock.module.ModuleMock.SingleParam with arguments ${cyan("equalTo(3)")}\n""")
   )
 
   val mock3 = zio.test.test("Extra calls") {
@@ -208,7 +208,7 @@ object ReportingTestUtils {
 
   val mock3Expected = Vector(
     expectedFailure("Extra calls"),
-    withOffset(2)(s"${red("- unexpected call to zio.test.mock.ModuleMock.ManyParams with arguments")}\n"),
+    withOffset(2)(s"${red("- unexpected call to zio.test.mock.module.ModuleMock.ManyParams with arguments")}\n"),
     withOffset(4)(s"${cyan("(2,3,4)")}\n")
   )
