+++ test/shared/src/main/scala-2.x/zio/test/mock/MockableMacro.scala
@@ -155,7 +155,7 @@ private[mock] object MockableMacro {
 
           ..$tags
 
-          private lazy val mock: _root_.zio.ZLayer[_root_.zio.Has[_root_.zio.test.mock.Mock], Nothing, $envType] =
+          private lazy val mock: _root_.zio.URLayer[_root_.zio.test.mock.MockRuntime, $envType] =
             _root_.zio.ZLayer.fromService(mock =>
               new $serviceType {
                 ..$mocks
