+++ test/shared/src/main/scala/zio/test/mock/package.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2017-2020 John A. De Goes and the ZIO Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package zio.test
+
+import zio.Has
+
+/**
+ * _ZIO Mock_ is a library for mocking collaborators.
+ *
+ * The library treats every expectation as an ordinary immutable value, enabling
+ * creation of complex expectations by means of composition. Thanks to tight
+ * integration with ZIO Test all the expectations will be checked and any mismatch
+ * properly reported failing the test.
+ */
+package object mock {
+
+  type MockRuntime = Has[MockRuntime.Service]
+}
